Use fixed org.opendaylight.daexim instead <groupId>${project.groupId}
to work around what appears to be a bug in the Maven Eclipse integration
M2E, which struggles with (fails to do) "workspace resolution" of
<scope>import *-artifacts projects using ${} variables.

With this change, a Import Projects of daexim into Eclipse, e.g. by
https://github.com/vorburger/opendaylight-eclipse-setup (or normal
Eclipse), "just works" out of the box; that's always nice, if it can be
easily achieved.  Without this, you have red, and have to first do a CLI
mvn install.

I suspect that maybe it just happens to work in other ODL projects using
<groupId>${project.groupId} because it downloads the *-artifacts
artifacts from nexus.opendaylight.org when encountering a <scope>import
(not sure, not tested; but that would explain why I haven't run into
this problem before - or maybe no other ODL projet does this).  [This
would mean that when you change dependencies in a <scope>import
*-artifacts used by <groupId>${project.groupId}, then it may / probably
doesn't actually do workspace resolution - you have to mvn install.]

https://bugs.eclipse.org/bugs/show_bug.cgi?id=514011 raised upstream to
perhaps one day get this fixed in M2E - in the mean time, simplest is to
just work around like this.

Change-Id: I68bff1bc3c3c8de496a6b0d34789c9b71682db7e
Signed-off-by: Michael Vorburger <vorburger@redhat.com>