Added LOG.error() in all catch of the RPC operations/Import/Util
I'm not extremely familiar with ODL's RPC mechanism yet, so I'm not 100%
sure if this could not lead to "double exception logging" - if whatever
subsystem calls the RPC correctly deals with the future (but they all?),
then this may not strictly necessary... but it shouldn't hurt either?
Better one error log too much than missing one.  

In the tests alone this should be useful to see why things fail.

Change-Id: I5b2c77037c3c00c7cf5671ebdf3b789332e67b47
Signed-off-by: Michael Vorburger <vorburger@redhat.com>