Merge odl-daexim-depends feature into odl-daexim-all
There is no point in having a separate feature for dependencies... a
feature should be a "functional unit" that an end-user operator may want
to install, like an application or a package from something like a
package manager on Linux like RPM or DEB.  Splitting the technical
dependencies into a separate feature has little value - nobody would
ever want to install just those and not the "actual feature".  

Other ODL projects typically don't do this either, so this is also more
consistent, and leads to 1 feature less overall (there are already way
too many anyway).

This will also (OK only slightly) simplify the migration to Karaf 4.

Change-Id: I603ee25c1627622b54714e4c8263834a001be499
Signed-off-by: Michael Vorburger <vorburger@redhat.com>